+++ SmartMealFinderB/src/main/java/hu/project/smartmealfinderb/Request/DailyProgressPostReq.java
@@ -6,11 +6,11 @@ import lombok.Data;
 @Data
 public class DailyProgressPostReq {
 
-    private float weight;
-    private int caloriesConsumed;
-    private int proteinConsumed;
-    private int carbsConsumed;
-    private int fatsConsumed;
+    private double weight;
+    private double caloriesConsumed;
+    private double proteinConsumed;
+    private double carbsConsumed;
+    private double fatsConsumed;
 
     private String comment;
 }
